Make spec types uninhabited to prevent creating objects
Spec types are used in a namespace-like way, and they are not meant to
be used to create values of the spec types.
Any types can used in this way, but to prevent accidentally creating
meaningless value, spec types are better to be uninhabited types.

When `!` (never type) is stabilized, `struct FooSpec(!);` is recommended
because it makes clear that a value of the type cannot be created.
However, it is currently unstable, so use `enum FooSpec {}` for now.